requirements.txt
kornia==0.7.0
numpy==1.22.0
pandas==2.0.2
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
